+++ index.html
@@ -8,18 +8,6 @@
   </head>
   <body>
     NoneOS
-    <script>
-      if ("serviceWorker" in navigator) {
-        navigator.serviceWorker
-          .register("base-agent.js", { type: "module" })
-          .then((reg) => {
-            setTimeout(() => {
-              // 一小时后自动更新
-              reg.update();
-            }, 60 * 60 * 1000);
-          })
-          .catch((err) => {});
-      }
-    </script>
+    <script src="import-proxy.js"></script>
   </body>
 </html>
